requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# WebSocket and async
websockets>=12.0  # Requires additional_headers instead of extra_headers
python-multipart==0.0.6

# Twilio for telephony
twilio==8.10.0

# Supabase for database
supabase==2.3.0

# Monitoring
prometheus-client==0.20.0

# Redis (optional, for caching)
redis==5.0.1

# HTTP requests (compatible with supabase <0.25.0)
httpx==0.24.1
